feat(PrincipalCliente): classe implementada

+++ src/main/java/com/carlosribeiro/service/ClienteService.java
@@ -1,15 +1,19 @@
 package com.carlosribeiro.service;
 
 import com.carlosribeiro.dao.ClienteDAO;
+import com.carlosribeiro.exception.ClienteAssociadoException;
+import com.carlosribeiro.exception.EmUso;
 import com.carlosribeiro.exception.EntidadeNaoEncontradaException;
 import com.carlosribeiro.model.Cliente;
 import com.carlosribeiro.util.FabricaDeDaos;
+
+import java.util.List;
 
 public class ClienteService {
 
     ClienteDAO clienteDAO = FabricaDeDaos.getDAO(ClienteDAO.class);
 
-    //todo adicionar regras de negócio
+
     public Cliente incluir(Cliente cliente) {
         clienteDAO.incluir(cliente);
         return cliente;
@@ -18,8 +22,14 @@ public class ClienteService {
     //todo regras de négocio
     public Cliente remover(int id ) {
         Cliente cliente = recuperarPorId(id) ;
-        cliente = clienteDAO.remover(id) ;
-        return cliente;
+        //remover se não tiver fatura ou pedidos ;
+        if( cliente.getPedidos().isEmpty() && cliente.getFaturas().isEmpty() ){
+            cliente = clienteDAO.remover(id) ;
+            return cliente;
+        }else{
+            throw new ClienteAssociadoException( "Cliente Associado a pedido ou fatura" ) ;
+        }
+
     }
 
 
@@ -29,6 +39,11 @@ public class ClienteService {
             throw new EntidadeNaoEncontradaException( "Cliente inexistente" ) ;
         }
         return cliente ;
+    }
+
+
+    public List<Cliente> recuperarTodoss(){
+        return clienteDAO.recuperarTodos();
     }
